==> 1.ODI Data Segmentation.ipynb <==
import os
import pandas as pd
import numpy as np
import json

odi_dir='E:/Cricket Prediction Project/odis_json'
t20_dir='E:/Cricket Prediction Project/t20s_json'

the_10=['England','Australia','New Zealand','South Africa','West Indies','India','Pakistan','Afghanistan','Bangladesh','Sri Lanka']

name1="1st_Innings.csv"
name2="2nd_Innings.csv"
dir_path_men="E:/Cricket Prediction Project/ODI/ODI_MEN"
dir_path_women="E:/Cricket Prediction Project/ODI/ODI_WOMEN"

for filename in os.listdir(odi_dir):
    json_file=open(os.path.join(odi_dir,filename))
    file=json.load(json_file)
    meta_data = {
        "gender":file["info"]["gender"],
        "teams":file["info"]["teams"],
        "venue":file["info"]["venue"]
    }
    data_frame_first_innings={
        "delivery":[],
        "batter":[],
        "non_striker":[],
        "bowler":[],
        "run_hit":[],
        "extras":[],
        "type_of_extras":[],
        "total":[],
        "wicket":[],
        "player_dismissed":[],
        "mode_of_dismissal":[]
    }
    if(len(file["innings"])>=1):
        for i in file["innings"][0]["overs"]:
            c=1
            for j in i["deliveries"]:
                ball_no=str(i["over"])+"."+str(c)
                batter=j["batter"]
                bowler=j["bowler"]
                non_striker=j["non_striker"]
                run_hit=j["runs"]["batter"]
                extras=j["runs"]["extras"]
                type_of_extras="null"
                total=j["runs"]["total"]
                wickets=0
                player_dismissed="null"
                mode_of_dismissal="null"
                if "wickets" in j:
                    wickets=1
                    player_dismissed=j["wickets"][0]["player_out"]
                    mode_of_dismissal=j["wickets"][0]["kind"]
                if "extras" in j:
                    for k in j["extras"]:
                        type_of_extras=str(k)+" "
                data_frame_first_innings["delivery"].append(ball_no)
                data_frame_first_innings["batter"].append(batter)
                data_frame_first_innings["non_striker"].append(non_striker)
                data_frame_first_innings["bowler"].append(bowler)
                data_frame_first_innings["run_hit"].append(run_hit)
                data_frame_first_innings["extras"].append(extras)
                data_frame_first_innings["type_of_extras"].append(type_of_extras)
                data_frame_first_innings["total"].append(total)
                data_frame_first_innings["wicket"].append(wickets)
                data_frame_first_innings["player_dismissed"].append(player_dismissed)
                data_frame_first_innings["mode_of_dismissal"].append(mode_of_dismissal)
                c+=1
            
    data_frame_second_innings={
        "delivery":[],
        "batter":[],
        "non_striker":[],
        "bowler":[],
        "run_hit":[],
        "extras":[],
        "type_of_extras":[],
        "total":[],
        "wicket":[],
        "player_dismissed":[],
        "mode_of_dismissal":[]
        }
    if(len(file["innings"])>=2):
        for i in file["innings"][1]["overs"]:
            c=1
            for j in i["deliveries"]:
                ball_no=str(i["over"])+"."+str(c)
                batter=j["batter"]
                bowler=j["bowler"]
                non_striker=j["non_striker"]
                run_hit=j["runs"]["batter"]
                extras=j["runs"]["extras"]
                type_of_extras="null"
                total=j["runs"]["total"]
                wickets=0
                player_dismissed="null"
                mode_of_dismissal="null"
                if "wickets" in j:
                    wickets=1
                    player_dismissed=j["wickets"][0]["player_out"]
                    mode_of_dismissal=j["wickets"][0]["kind"]
                if "extras" in j:
                    for k in j["extras"]:
                        type_of_extras=str(k)+" "
                data_frame_second_innings["delivery"].append(ball_no)
                data_frame_second_innings["batter"].append(batter)
                data_frame_second_innings["non_striker"].append(non_striker)
                data_frame_second_innings["bowler"].append(bowler)
                data_frame_second_innings["run_hit"].append(run_hit)
                data_frame_second_innings["extras"].append(extras)
                data_frame_second_innings["type_of_extras"].append(type_of_extras)
                data_frame_second_innings["total"].append(total)
                data_frame_second_innings["wicket"].append(wickets)
                data_frame_second_innings["player_dismissed"].append(player_dismissed)
                data_frame_second_innings["mode_of_dismissal"].append(mode_of_dismissal)
                c+=1
    data1=pd.DataFrame(data_frame_first_innings)
    data2=pd.DataFrame(data_frame_second_innings)
    dir_name=str(meta_data["teams"][0]) + " vs " + str(meta_data["teams"][1]) + " " + str(filename)
    if meta_data["gender"]=="male" :
        path=os.path.join(dir_path_men,dir_name)
    else:
        path=os.path.join(dir_path_women,dir_name)
    try:
        os.mkdir(path)
        data1.to_csv(os.path.join(path,name1))
        data2.to_csv(os.path.join(path,name2))
        data=json.dumps(meta_data)
        with open(os.path.join(path,"meta_data.json"), "w") as outfile:
            outfile.write(data)
    except:
        print(dir_name+" exists twice")

Testing the code while developement

json_file=open("odis_json/64846.json")
file=json.load(json_file)
data_frame_first_innings={
        "delivery":[],
        "batter":[],
        "non_striker":[],
        "bowler":[],
        "run_hit":[],
        "extras":[],
        "type_of_extras":[],
        "total":[],
        "wicket":[],
        "player_dismissed":[],
        "mode_of_dismissal":[]
    }
for i in file["innings"][0]["overs"]:
    k=i["over"]
    c=1
    for j in i["deliveries"]:
        ball_no=str(k)+"."+str(c)
        batter=j["batter"]
        bowler=j["bowler"]
        non_striker=j["non_striker"]
        run_hit=j["runs"]["batter"]
        extras=j["runs"]["extras"]
        type_of_extras="null"
        total=j["runs"]["total"]
        wickets=0
        player_dismissed="null"
        mode_of_dismissal="null"
        if "wickets" in j:
            wickets=1
            player_dismissed=j["wickets"][0]["player_out"]
            mode_of_dismissal=j["wickets"][0]["kind"]
        if "extras" in j:
            for k in j["extras"]:
                type_of_extras+=str(k)+" "
        data_frame_first_innings["delivery"].append(ball_no)
        data_frame_first_innings["batter"].append(batter)
        data_frame_first_innings["non_striker"].append(bowler)
        data_frame_first_innings["bowler"].append(non_striker)
        data_frame_first_innings["run_hit"].append(run_hit)
        data_frame_first_innings["extras"].append(extras)
        data_frame_first_innings["type_of_extras"].append(type_of_extras)
        data_frame_first_innings["total"].append(total)
        data_frame_first_innings["wicket"].append(wickets)
        data_frame_first_innings["player_dismissed"].append(player_dismissed)
        data_frame_first_innings["mode_of_dismissal"].append(mode_of_dismissal)
        c+=1

Saving the Data

data=pd.DataFrame(data_frame_first_innings)

data.head(40)

name="file"+".csv"
k="D:/Cricket Prediction Project/ODI"
data.to_csv(os.path.join(k,name))


ds={0:1,5:5}
with open('convert.txt', 'w') as convert_file:
     convert_file.write(json.dumps(ds))
