==> 4.Batter Bowler Allrounder Classification.ipynb <==
import json 
import os
import pandas as pd

df=pd.read_csv("E:/Cricket Prediction Project/Colab Notebooks/Players.csv", index_col=0)

df

df['Bowling'].unique()

types_of_batters=["Left Handed Bat","Right Handed Bat","Right hand Bat","Left hand Bat"]


types_of_bowlers=['Right-arm legbreak','Right-arm offbreak', 'Left-arm orthodox','Right-arm fast-medium','Left-arm fast-medium',
       'Right-arm fast','Left-arm fast','Right-arm medium','Left-arm chinaman','Right arm Fast medium','Right arm Medium fast',
        'Left-arm medium', 'Right arm Medium', 'Legbreak','Right arm Offbreak','Slow Left arm Orthodox','Left arm Fast medium',
                  'Slow Left arm Orthodox']

df["Position"].unique()

batters=df.loc[df["Position"].isin(['Batsman','WK-Batsman','Wicketkeeper Batter', 'Middle order Batter', 'Top order Batter','Opening Batter'])]

batters.index=list(range(638))

bowlers=df.loc[df["Position"].isin(["Bowler", "Bowling Allrounder"])]

bowlers.index=list(range(578))

allrounders=df.loc[df["Position"].isin(["Allrounder", "Batting Allrounder"])]

allrounders.index=list(range(283))

batters.to_csv("Batters.csv")
bowlers.to_csv("Bowlers.csv")
allrounders.to_csv("Allrounders.csv")
