==> 6.Batter Features.ipynb <==
import os
import json
import pandas as pd

data = {
    "Name": [],
    "Right arm Fast": [],
    "Right arm Medium": [],
    "Left arm Fast": [],
    "Left arm Medium": [],
    "Right arm off-spinner": [],
    "Right arm Legbreak": [],
    "Left arm Orthodox": [],
    "Left arm Chinaman": [],
    "Hitting Score": [],
    "Anchoring Score": []
}

players='E:\Cricket Prediction Project\PLAYERS'

def fetch_data(feature,df):
    runs_scored=sum(list(df.query('bowler_type_tuple==@feature')["runs_scored"]))
    balls_faced=sum(list(df.query('bowler_type_tuple==@feature')["balls_faced"]))
    got_him_out=sum(list(df.query('bowler_type_tuple==@feature')["got_him_out"]))
    return str([runs_scored,balls_faced,got_him_out])

for filename in os.listdir(players):
    df=pd.read_csv(os.path.join(players,filename))
    player_name=filename[:-4]
    data["Name"].append(player_name)
    runs_scored=sum(list(df.query('bowler_type_tuple=="[0, 1, 0]" or bowler_type_tuple=="[0, 1, 1]"')["runs_scored"]))
    balls_faced=sum(list(df.query('bowler_type_tuple=="[0, 1, 0]" or bowler_type_tuple=="[0, 1, 1]"')["balls_faced"]))
    got_him_out=sum(list(df.query('bowler_type_tuple=="[0, 1, 0]" or bowler_type_tuple=="[0, 1, 1]"')["got_him_out"]))
    data["Right arm Fast"].append(str([runs_scored,balls_faced,got_him_out]))
    data["Right arm Medium"].append(fetch_data("[0, 1, 2]", df))
    runs_scored=sum(list(df.query('bowler_type_tuple=="[0, 0, 0]" or bowler_type_tuple=="[0, 0, 1]"')["runs_scored"]))
    balls_faced=sum(list(df.query('bowler_type_tuple=="[0, 0, 0]" or bowler_type_tuple=="[0, 0, 1]"')["balls_faced"]))
    got_him_out=sum(list(df.query('bowler_type_tuple=="[0, 0, 0]" or bowler_type_tuple=="[0, 0, 1]"')["got_him_out"]))
    data["Left arm Fast"].append(str([runs_scored,balls_faced,got_him_out]))
    data["Left arm Medium"].append(fetch_data("[0, 0, 2]", df))
    data["Right arm off-spinner"].append(fetch_data("[1, 1, 0]", df))
    data["Right arm Legbreak"].append(fetch_data("[1, 1, 1]", df))
    data["Left arm Orthodox"].append(fetch_data("[1, 0, 1]", df))
    data["Left arm Chinaman"].append(fetch_data("[1, 0, 0]", df))
    data["Hitting Score"].append("To be calculated")
    data["Anchoring Score"].append("To be calculated")
data=pd.DataFrame(data)
data.to_csv("E:\Cricket Prediction Project\Condensed.csv")
