==> 8.Finding Anchoring and Hitting Score.ipynb <==
import os
import pandas as pd

Now, to calculate Anchoring Score methods that come to mind are -
i) Boundaries per ball
ii) Normal Strike Rate
iii) Percentage Runs from Boundaries.
iv) Weighted average of Dot balls played and Boundary scoring balls

data={
    "Name":[],
    "Playing Style Score":[]
}

for player in os.listdir("E:/Cricket Prediction Project/PLAYERS"):
    df=pd.read_csv("E:/Cricket Prediction Project/PLAYERS/"+player) 
    if sum(df["balls_faced"]) == 0 or sum(df["runs_scored"])==0:
        continue
    data["Name"].append(player[:-4])
    data["Playing Style Score"].append((sum(df["fours"])*4+sum(df["sixes"])*6)/sum(df["runs_scored"]) * 100)                                        

df=pd.DataFrame(data)

df_condensed = pd.read_csv("E:/Cricket Prediction Project/Condensed.csv", index_col=0)

ls=list(df["Name"])

ls

df_sorted=df_condensed.query('Name in @ls')

df_sorted

df_sorted=df_sorted.reset_index()
df=df.reset_index()

k=pd.concat([df_sorted, df.drop('Name', axis='columns')], axis=1, join='inner')

k.columns

k=k.drop('Hitting Score', axis='columns')
k=k.drop('Anchoring Score', axis='columns')
k=k.drop('index', axis='columns')

k.to_csv('E:/Cricket Prediction Project/Final Data.csv')

k
